Broad clean up
Listen on default port 3000 if PORT_... is not defined.  Allow a
lowercase atom to specify the service name (more natural this way).
Also allow the service to be started without using http_handler/1.  That
makes it possible to get access directly to all requests as they come
in.

Improve documentation too.